[HAPI-184] Create the CRUD class instance
## [HAPI-184] Description
Created a class for the CRUD operations
