+++ src/collections/Class/_globals.class.js
@@ -1,5 +1,3 @@
-const CRUD = require('4hands-api/src/services/database/crud');
-
 class GlobalClass {
     initialize() {
         try {
@@ -12,32 +10,6 @@ class GlobalClass {
                 return this;
             }
         } catch(err) {
-            throw logError(err);
-        }
-    }
-
-    async defaultPopulate() {
-        try {
-            const docQuery = CRUD.getDoc({collectionName: this.collection.collectionName, filter: this.id });
-            
-            if (docQuery.defaultPopulate) {
-                const docPopulated = await docQuery.defaultPopulate();
-                if (docPopulated.error) {
-                    throw docPopulated;
-                }
-
-                if (docPopulated) {
-                    return docPopulated;
-                } else {
-                    return null;
-                }
-            } else {
-                return logError({
-                    name: 'MONGOOSE-QUERY-NOT-FOUND',
-                    message: `The mongoose custom query "defaultPopulate" don't exist!`
-                });
-            }
-        } catch (err) {
             throw logError(err);
         }
     }
