+++ src/helpers/database/index.js
@@ -1,11 +1,7 @@
 const dbHelpers = require('./dbHelpers');
-const queries = require('4hands-api/src/collections/queries/_globals.query');
-const events = require('4hands-api/src/collections/events/_globals.event');
 const relationalFields = require('./relationalFields');
 
 module.exports = {
     dbHelpers,
-    queries,
-    events,
     relationalFields
 };
