+++ src/models/collections/SafeValue.js
@@ -64,7 +64,7 @@ class SafeValue extends _Global {
      * @static
      */
     static async createEncrypt(rawValue) {
-        const CRUD = require('4hands-api/src/services/database/crud');
+        const CRUD = require('4hands-api/src/services/Database/CRUD');
 
         return await CRUD.create('safe_values', {
             type: 'encrypt',
